updated unit tests toolkit initialize logic
